extracted intellisense logic
